Update v0.3 - Guilty

- Implantation du Kolizeum
- Amélioration du temps de démarrage de l'émulateur
- Résolution du problème lié à l'xp donnée à la guilde
- Retrait des boucliers lors de la désactivation de l'alignement
- Correction de crashs émulateur
- Correction des loots du percepteur
- Résolution du bug des tours infini des bonbons.
- Quelques corrections sur le NPC échangeur de Dofus légendaire
- Possibilité d'insérer les caractéristiques des objets dans le chat
- Débug du sort des Pandawa, lait de bambou
- Les points de statistiques sont maintenant correctement ajoutés
- Correction du bug combat lors du kick d'un joueur en combat alors que la team est complète

+++ Server/Stump.Server.WorldServer/Game/Exchanges/Trades/Npcs/NpcTrade.cs
@@ -78,12 +78,12 @@ namespace Stump.Server.WorldServer.Game.Exchanges.Trades.Npcs
             foreach (var tradeItem in FirstTrader.Items)
             {
                 var item = FirstTrader.Character.Inventory.TryGetItem(tradeItem.Guid);
-                    FirstTrader.Character.Inventory.RemoveItem(item, tradeItem.Stack);
+                    FirstTrader.Character.Inventory.RemoveItem(item, (int)tradeItem.Stack);
             }
 
             foreach (var tradeItem in SecondTrader.Items)
             {
-                FirstTrader.Character.Inventory.AddItem(tradeItem.Template, tradeItem.Stack);
+                FirstTrader.Character.Inventory.AddItem(tradeItem.Template, (int)tradeItem.Stack);
             }
 
             InventoryHandler.SendInventoryWeightMessage(FirstTrader.Character.Client);
